# crates/arco-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arco API",
    "description": "Arco catalog REST API (M4/M5)",
    "contact": {
      "name": "Daxis",
      "email": "engineering@daxis.io"
    },
    "license": {
      "name": "Apache-2.0 OR MIT",
      "identifier": "Apache-2.0 OR MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/namespaces": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "List all namespaces.",
        "description": "GET /api/v1/namespaces",
        "operationId": "list_namespaces",
        "responses": {
          "200": {
            "description": "Namespaces listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNamespacesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": "Create a namespace.",
        "description": "POST /api/v1/namespaces",
        "operationId": "create_namespace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNamespaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Namespace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{name}": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "Get a namespace by name.",
        "description": "GET /api/v1/namespaces/{name}",
        "operationId": "get_namespace",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Namespace name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namespace found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "namespaces"
        ],
        "summary": "Delete a namespace.",
        "description": "DELETE /api/v1/namespaces/{name}",
        "operationId": "delete_namespace",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Namespace name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Namespace deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Standard JSON error response body.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message (safe for clients)."
          },
          "requestId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional request ID for correlation."
          }
        }
      },
...
      "CreateNamespaceRequest": {
        "type": "object",
        "description": "Request to create a namespace.",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description."
          },
          "name": {
            "type": "string",
            "description": "Namespace name (must be unique within workspace)."
          }
        }
      },
...
      "ListNamespacesResponse": {
        "type": "object",
        "description": "List namespaces response.",
        "required": [
          "namespaces"
        ],
        "properties": {
          "namespaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamespaceResponse"
            },
            "description": "List of namespaces."
          }
        }
      },
...
      "NamespaceResponse": {
        "type": "object",
        "description": "Namespace response.",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "Creation timestamp (ISO 8601)."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description."
          },
          "id": {
            "type": "string",
            "description": "Namespace ID."
          },
          "name": {
            "type": "string",
            "description": "Namespace name."
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp (ISO 8601)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "taskAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}